4/7

+++ Backend/src/migrations/20250515043514-create-company-review.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       rating: {
-        type: Sequelize.TINYINT
+        type: Sequelize.INTEGER
       },
       review: {
         type: Sequelize.TEXT
